use REST API for repository related content
