+++ src/data/queries/repositoryRelatedContentQuery.ts
@@ -1,0 +1,6 @@
+import { useSearchDoiFacetsQuery } from './searchDoiFacetsQuery'
+
+export function useRepositoryRelatedContent(id: string) {
+  const vars = { clientId: id }
+  return useSearchDoiFacetsQuery(vars)
+}
